+++ docs/changes.rst
@@ -4,7 +4,9 @@ Changelog
 
 Version 2.0.0 (Future Release)
 ==============================
- - Support Graph API version 2.6 and 2.7.
+ - Add support for Graph API versions 2.6 and 2.7.
+ - Remove support for Graph API version 2.0 and FQL.
+ - Change default Graph API version to 2.1.
  - Allow offline generation of application access tokens.
 
 Version 1.0.0 (2016-04-01)